requirements.txt
Pillow==10.2.0
pyinstaller==6.4.0